support: filter out after run

+++ src/Point/Point2d.hh
@@ -2,14 +2,13 @@
 #define VLL_DEEPARC_POINT2D_H_
 class Point2d{
     public:
-        Point2d(int x, int y){
+        Point2d(double x, double y){
             this->x_ = x;
             this->y_ = y;
         }
-        int x(){return x_;}
-        int y(){return y_;}
-
+        double x(){return x_;}
+        double y(){return y_;}
     private:
-    int x_, y_;
+        double x_, y_;
 };
 #endif
